Add and integrate sdspi and fatfs sources
sdspi is a hardware module to talk to sd cards over SPI.
It also comes with software drivers that had to be slightly corrected
and adapted to the neorv32 core.

FATFS is a fat filesystem driver.

Both sdspi and fatfs are included as git submodules.
The fatfs repo used for this is not maintained by the original author
of fatfs but tracks the fatfs sources published on the project website.

The sdspi submodule uses a fork with the necessary changes applied
for compatibility with the neorv32.

The sdspi hdl files are added to the project.tcl file by sourcing the
add_spi.tcl script. The sdspi module was connected to the neorv32 using
its wishbone bus.

Finally, several example and test programs were added to
saspit/sw/programs/.